refactor: small file and folder renames
