+++ elasticai/creator/translatable_modules/vhdl/fp_linear_1d/design.py
