+++ elasticai/creator/translatable_modules/vhdl/lstm/lstm.py
@@ -6,7 +6,7 @@ import torch.nn
 from elasticai.creator.hdl.design_base.design import Design
 from elasticai.creator.nn.lstm import FixedPointLSTMWithHardActivations as _nnLSTM
 from elasticai.creator.translatable_modules.vhdl.fp_linear_1d import FPLinear1d
-from elasticai.creator.translatable_modules.vhdl.fp_linear_1d.fp_linear_1d_hw_design import (
+from elasticai.creator.translatable_modules.vhdl.fp_linear_1d.design import (
     FPLinear1d as _FPLinear1dDesign,
 )
 from elasticai.creator.translatable_modules.vhdl.lstm.design import LSTMNetworkDesign
